Use independent GWL values on the X axis
This version does not use decades as parameter to select both the global
and regional warming levels. Instead, GWL values are chosen as an
independent variable and the corresponding regional warming is computed
from the period when the GWL is reached for the first time.

Notebook text still need to be updated.